pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pyising"
version = "0.2.0"
description = "2D Ising model simulation with optional GPU acceleration via CuPy"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"

dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "matplotlib>=3.7",
    "tqdm>=4.65",
]

[project.optional-dependencies]
gpu = [
    # Install the variant that matches your CUDA version:
    #   cupy-cuda11x  →  CUDA 11.x
    #   cupy-cuda12x  →  CUDA 12.x  (RTX 3090 default)
    "cupy-cuda12x>=12.0",
]
video = [
    "opencv-python>=4.8",
]
app = [
    "marimo>=0.9.0",
]
all = [
    "cupy-cuda12x>=12.0",
    "opencv-python>=4.8",
    "marimo>=0.9.0",
]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.1",
    "marimo>=0.9.0",
]

[project.scripts]
pyising-info = "pyising.hardware:print_summary"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
